requirements.txt
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
langchain-openai
langchain-community
streamlit
PyMuPDF

